+++ app/src/main/java/com/monke/monkeybook/model/content/BookContent.java
@@ -8,8 +8,10 @@ import com.monke.monkeybook.dao.ChapterListBeanDao;
 import com.monke.monkeybook.dao.DbHelper;
 import com.monke.monkeybook.help.RxBusTag;
 import com.monke.monkeybook.model.AnalyzeRule.AnalyzeElement;
+import com.monke.monkeybook.model.AnalyzeRule.AnalyzeJson;
 import com.monke.monkeybook.model.ErrorAnalyContentManager;
 
+import org.json.JSONObject;
 import org.jsoup.Jsoup;
 import org.jsoup.nodes.Document;
 
@@ -33,9 +35,15 @@ public class BookContent {
             bookContentBean.setDurChapterUrl(durChapterUrl);
             bookContentBean.setTag(tag);
             try {
-                Document doc = Jsoup.parse(s);
-                AnalyzeElement analyzeElement = new AnalyzeElement(doc, durChapterUrl);
-                bookContentBean.setDurChapterContent(analyzeElement.getResult(bookSourceBean.getRuleBookContent()));
+                if (bookSourceBean.getRuleBookContent().contains("JSON")) {
+                    JSONObject jsonObject = new JSONObject(s);
+                    AnalyzeJson analyzeJson = new AnalyzeJson(jsonObject);
+                    bookContentBean.setDurChapterContent(analyzeJson.getResult(bookSourceBean.getRuleBookContent()));
+                } else {
+                    Document doc = Jsoup.parse(s);
+                    AnalyzeElement analyzeElement = new AnalyzeElement(doc, durChapterUrl);
+                    bookContentBean.setDurChapterContent(analyzeElement.getResult(bookSourceBean.getRuleBookContent()));
+                }
                 bookContentBean.setRight(true);
             } catch (Exception ex) {
                 ex.printStackTrace();
